fix: add arrayUtils, fix playlist-shuffle cmd
