+++ src/commands/stream/youtubeSearch.ts
@@ -2,11 +2,12 @@ import LocalServer from '../../utils/lib/LocalServer';
 import { Message, MessageReaction, User } from 'discord.js';
 import ytdl from 'ytdl-core-discord';
 import { playLinkToVC } from './stream';
-import { adjustQueueForPlayNow, createQueueItem } from '../../utils/utils';
+import { createQueueItem } from '../../utils/utils';
 import { botID, StreamType } from '../../utils/lib/constants';
 import { updateActiveEmbed } from '../../utils/embed';
 import reactions from '../../utils/lib/reactions';
 import EmbedBuilderLocal from '../../utils/lib/EmbedBuilderLocal';
+import { adjustQueueForPlayNow } from '../../utils/arrayUtils';
 
 const ytsr = require('ytsr');
